# docs/reference/openapi-compliance.yaml
openapi: 3.1.0
info:
  title: Luqen Compliance Service
  description: >-
    Accessibility compliance rule engine — check WCAG issues against 60+
    country-specific legal requirements. Provides CRUD for jurisdictions,
    regulations, and requirements, a compliance checker, update proposal
    workflow, monitored sources, webhooks, user/client management, and OAuth
    token issuance.
  version: 0.1.0
  license:
    name: MIT
servers:
  - url: http://localhost:4000
    description: Local development
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/compliance/check:
    post:
      operationId: checkCompliance
      tags: [compliance]
      summary: Check WCAG issues against jurisdictional requirements
      description: >-
        Accepts a list of jurisdictions and accessibility issues, returns a
        compliance matrix with per-jurisdiction and per-regulation results plus
        annotated issues. Results may be cached (Redis) — check the `X-Cache`
        response header for `HIT` or `MISS`.
      parameters:
        - $ref: "#/components/parameters/XOrgId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ComplianceCheckRequest"
            example:
              jurisdictions: [eu, us-ca]
              issues:
                - code: WCAG2AA.Principle1.Guideline1_1.1_1_1.H37
                  type: error
                  message: Img element missing an alt attribute
                  selector: "#hero img"
                  context: "<img src=\"hero.jpg\">"
              includeOptional: false
              sectors: [public]
      responses:
        "200":
          description: Compliance check result
          headers:
            X-Cache:
              description: Cache status — `HIT` if served from cache, `MISS` if freshly computed.
              schema:
                type: string
                enum: [HIT, MISS]
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ComplianceCheckResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"

  # ────────────────────────────────────────────────────────────────────────────
  # Updates (proposals)
  # ────────────────────────────────────────────────────────────────────────────
...
  /api/v1/sources:
    get:
      operationId: listSources
      tags: [sources]
      summary: List monitored sources
      parameters:
        - $ref: "#/components/parameters/XOrgId"
      responses:
        "200":
          description: Source list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MonitoredSource"
        "500":
          $ref: "#/components/responses/InternalError"

    post:
      operationId: createSource
      tags: [sources]
      summary: Create a monitored source
      description: Requires `admin` scope.
      parameters:
        - $ref: "#/components/parameters/XOrgId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSourceInput"
      responses:
        "201":
          description: Source created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MonitoredSource"
        "400":
          $ref: "#/components/responses/BadRequest"
  /api/v1/sources/{id}:
    delete:
      operationId: deleteSource
      tags: [sources]
      summary: Delete a monitored source
      description: Requires `admin` scope.
      parameters:
        - $ref: "#/components/parameters/ResourceId"
        - $ref: "#/components/parameters/XOrgId"
      responses:
        "204":
          description: Source deleted
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/sources/scan:
    post:
      operationId: scanSources
      tags: [sources]
      summary: Trigger a scan of all monitored sources
      description: >-
        Fetches each monitored source, computes a SHA-256 content hash, and
        creates update proposals for any sources whose content has changed.
        Requires `admin` scope.
      parameters:
        - $ref: "#/components/parameters/XOrgId"
      responses:
        "200":
          description: Scan results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScanSourcesResponse"
        "500":
          $ref: "#/components/responses/InternalError"

  # ────────────────────────────────────────────────────────────────────────────
  # Webhooks
  # ────────────────────────────────────────────────────────────────────────────
  /api/v1/webhooks:
    get:
      operationId: listWebhooks
      tags: [webhooks]
      summary: List webhooks
      description: Requires `admin` scope.
      parameters:
        - $ref: "#/components/parameters/XOrgId"
      responses:
        "200":
          description: Webhook list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Webhook"
        "500":
          $ref: "#/components/responses/InternalError"

    post:
      operationId: createWebhook
      tags: [webhooks]
      summary: Create a webhook
      description: Requires `admin` scope.
      parameters:
        - $ref: "#/components/parameters/XOrgId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateWebhookInput"
      responses:
        "201":
          description: Webhook created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Webhook"
        "400":
          $ref: "#/components/responses/BadRequest"
  /api/v1/webhooks/{id}:
    delete:
      operationId: deleteWebhook
      tags: [webhooks]
      summary: Delete a webhook
      description: Requires `admin` scope.
      parameters:
        - $ref: "#/components/parameters/ResourceId"
        - $ref: "#/components/parameters/XOrgId"
      responses:
        "204":
          description: Webhook deleted
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/webhooks/{id}/test:
    post:
      operationId: testWebhook
      tags: [webhooks]
      summary: Send a test payload to a webhook
      description: >-
        Delivers a `webhook.test` event to the webhook URL and reports the
        HTTP status. Requires `admin` scope.
      parameters:
        - $ref: "#/components/parameters/ResourceId"
        - $ref: "#/components/parameters/XOrgId"
      responses:
        "200":
          description: Test delivery result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WebhookTestResponse"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"

  # ────────────────────────────────────────────────────────────────────────────
  # Users
  # ────────────────────────────────────────────────────────────────────────────
...
  /api/v1/seed:
    post:
      operationId: seedBaseline
      tags: [seed]
      summary: Load baseline seed data
      description: >-
        Loads the built-in baseline jurisdictions, regulations, and
        requirements. Requires `admin` scope.
      parameters:
        - $ref: "#/components/parameters/XOrgId"
      responses:
        "200":
          description: Seed completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SeedResponse"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/seed/status:
    get:
      operationId: getSeedStatus
      tags: [seed]
      summary: Get seed status
      description: Returns counts of seeded entities.
      parameters:
        - $ref: "#/components/parameters/XOrgId"
      responses:
        "200":
          description: Seed status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SeedStatus"
        "500":
          $ref: "#/components/responses/InternalError"

  # ────────────────────────────────────────────────────────────────────────────
  # Orgs
  # ────────────────────────────────────────────────────────────────────────────
  /api/v1/orgs/{id}/data:
    delete:
      operationId: deleteOrgData
      tags: [orgs]
      summary: Delete all data for an organization
      description: >-
        Removes all jurisdictions, regulations, requirements, proposals,
        sources, and webhooks belonging to the specified org. Cannot delete
        `system` org data. Requires `admin` scope.
      parameters:
        - $ref: "#/components/parameters/ResourceId"
        - $ref: "#/components/parameters/XOrgId"
      responses:
        "204":
          description: Org data deleted
        "400":
          description: Cannot delete system org data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          $ref: "#/components/responses/InternalError"

# ==============================================================================
# Components
# ==============================================================================
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    ResourceId:
      name: id
      in: path
      required: true
      schema:
        type: string
    XOrgId:
      name: X-Org-Id
      in: header
      description: >-
        Optional organization context. When set, scopes data operations to
        this organization.
      required: false
      schema:
        type: string
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    ErrorResponse:
      type: object
      required: [error, statusCode]
      properties:
        error:
          type: string
        statusCode:
          type: integer

    # ── Health ─────────────────────────────────────────────────────────────────
...
    ComplianceCheckRequest:
      type: object
      required: [jurisdictions, issues]
      properties:
        jurisdictions:
          type: array
          items:
            type: string
          minItems: 1
          description: Jurisdiction IDs to check against.
        issues:
          type: array
          items:
            $ref: "#/components/schemas/AccessibilityIssue"
          description: Accessibility issues to evaluate.
        includeOptional:
          type: boolean
          default: false
          description: Include optional requirements in the check.
        sectors:
          type: array
          items:
            type: string
          description: Filter regulations by sector.
    AccessibilityIssue:
      type: object
      required: [code, type, message, selector, context]
      properties:
        code:
          type: string
          description: Pa11y/WCAG issue code.
        type:
          type: string
          description: Issue severity (error, warning, notice).
        message:
          type: string
        selector:
          type: string
        context:
          type: string
          description: HTML context snippet.
        url:
          type: string
    ComplianceCheckResponse:
      type: object
      required: [matrix, annotatedIssues, summary]
      properties:
        matrix:
          type: object
          additionalProperties:
            $ref: "#/components/schemas/JurisdictionResult"
          description: Keyed by jurisdiction ID.
        annotatedIssues:
          type: array
          items:
            $ref: "#/components/schemas/AnnotatedIssue"
        summary:
          $ref: "#/components/schemas/ComplianceSummary"
    JurisdictionResult:
      type: object
      required:
        - jurisdictionId
        - jurisdictionName
        - status
        - mandatoryViolations
        - recommendedViolations
        - optionalViolations
        - regulations
      properties:
        jurisdictionId:
          type: string
        jurisdictionName:
          type: string
        status:
          type: string
          enum: [pass, fail]
        mandatoryViolations:
          type: integer
        recommendedViolations:
          type: integer
        optionalViolations:
          type: integer
        regulations:
          type: array
          items:
            $ref: "#/components/schemas/RegulationResult"
    RegulationResult:
      type: object
      required:
        - regulationId
        - regulationName
        - shortName
        - status
        - enforcementDate
        - scope
        - violations
      properties:
        regulationId:
          type: string
        regulationName:
          type: string
        shortName:
          type: string
        status:
          type: string
          enum: [pass, fail]
        enforcementDate:
          type: string
        scope:
          type: string
        violations:
          type: array
          items:
            type: object
            required: [wcagCriterion, obligation, issueCount]
            properties:
              wcagCriterion:
                type: string
              obligation:
                type: string
                enum: [mandatory, recommended, optional]
              issueCount:
                type: integer
    AnnotatedIssue:
      type: object
      required: [code, wcagCriterion, wcagLevel, originalIssue, regulations]
      properties:
        code:
          type: string
        wcagCriterion:
          type: string
        wcagLevel:
          type: string
        originalIssue:
          type: object
          additionalProperties: true
        regulations:
          type: array
          items:
            type: object
            required:
              - regulationId
              - regulationName
              - shortName
              - jurisdictionId
              - obligation
              - enforcementDate
            properties:
              regulationId:
                type: string
              regulationName:
                type: string
              shortName:
                type: string
              jurisdictionId:
                type: string
              obligation:
                type: string
                enum: [mandatory, recommended, optional]
              enforcementDate:
                type: string
    ComplianceSummary:
      type: object
      required:
        - totalJurisdictions
        - passing
        - failing
        - totalMandatoryViolations
        - totalOptionalViolations
      properties:
        totalJurisdictions:
          type: integer
        passing:
          type: integer
        failing:
          type: integer
        totalMandatoryViolations:
          type: integer
        totalOptionalViolations:
          type: integer

    # ── Update Proposal ────────────────────────────────────────────────────────
    UpdateProposal:
      type: object
      required: [id, source, detectedAt, type, summary, proposedChanges, status, createdAt]
      properties:
        id:
          type: string
        source:
          type: string
        detectedAt:
          type: string
          format: date-time
        type:
          type: string
          enum: [new_regulation, amendment, repeal, new_requirement, new_jurisdiction]
        affectedRegulationId:
          type: string
        affectedJurisdictionId:
          type: string
        summary:
          type: string
        proposedChanges:
          $ref: "#/components/schemas/ProposedChange"
        status:
          type: string
          enum: [pending, approved, rejected]
        reviewedBy:
          type: string
        reviewedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
    ProposedChange:
      type: object
      required: [action, entityType]
      properties:
        action:
          type: string
          enum: [create, update, delete]
        entityType:
          type: string
          enum: [jurisdiction, regulation, requirement]
        entityId:
          type: string
        before:
          type: object
          additionalProperties: true
        after:
          type: object
          additionalProperties: true
...
    MonitoredSource:
      type: object
      required: [id, name, url, type, schedule, createdAt]
      properties:
        id:
          type: string
        name:
          type: string
        url:
          type: string
          format: uri
        type:
          type: string
          enum: [html, rss, api]
        schedule:
          type: string
          enum: [daily, weekly, monthly]
        lastCheckedAt:
          type: string
          format: date-time
        lastContentHash:
          type: string
        createdAt:
          type: string
          format: date-time
    CreateSourceInput:
      type: object
      required: [name, url, type, schedule]
      properties:
        name:
          type: string
        url:
          type: string
          format: uri
        type:
          type: string
          enum: [html, rss, api]
        schedule:
          type: string
          enum: [daily, weekly, monthly]
    ScanSourcesResponse:
      type: object
      required: [scanned, proposalsCreated, proposals]
      properties:
        scanned:
          type: integer
          description: Number of sources scanned.
        proposalsCreated:
          type: integer
          description: Number of change-detected proposals created.
        proposals:
          type: array
          items:
            $ref: "#/components/schemas/UpdateProposal"

    # ── Webhook ────────────────────────────────────────────────────────────────
    Webhook:
      type: object
      required: [id, url, secret, events, active, createdAt]
      properties:
        id:
          type: string
        url:
          type: string
          format: uri
        secret:
          type: string
        events:
          type: array
          items:
            type: string
        active:
          type: boolean
        createdAt:
          type: string
          format: date-time
    CreateWebhookInput:
      type: object
      required: [url, secret, events]
      properties:
        url:
          type: string
          format: uri
        secret:
          type: string
        events:
          type: array
          items:
            type: string
    WebhookTestResponse:
      type: object
      properties:
        success:
          type: boolean
        statusCode:
          type: integer
          description: HTTP status from the webhook URL (only when success=true).
        error:
          type: string
          description: Error message (only when success=false).

    # ── User ───────────────────────────────────────────────────────────────────
...
    SeedResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
        jurisdictions:
          type: integer
        regulations:
          type: integer
        requirements:
          type: integer
    SeedStatus:
      type: object
      properties:
        jurisdictions:
          type: integer
        regulations:
          type: integer
        requirements:
          type: integer